+++ CHANGES.md
@@ -6,6 +6,7 @@ Elektron Change Log
 - [#15]: Configure mandatory email confirmation for new accounts.
 - [#16]: Default email backend will print the message to console.
 - [#3]: Fix error when trying to send email when signing up.
+- [#17]: Add test for email confirmation flow.
 
 0.21 (2019-01-08)
 -----------------